+++ core/main.ts
@@ -79,11 +79,11 @@ export * from "./src/ecs/components/Team";
 export * from "./src/ecs/actions/Boost";
 export * from "./src/ecs/actions/PlayerActions";
 export * from "./src/ecs/actions/Shoot";
+export * from "./src/ecs/actions/MovementActions";
 export * from "./src/ecs/actions/ToggleVisible";
 export * from "./src/ecs/actions/VehicleMovement";
 export * from "./src/ecs/actions/Wall";
-export * from "./src/ecs/actions/WASDMovement";
-export * from "./src/ecs/actions/ZombieMovement";
+export * from "./src/ecs/actions/WASDInputMap";
 export * from "./src/ecs/components/Controlling";
 export * from "./src/utils/MathUtils";
 export * from "./src/utils/PixiUtils";
